dm vdo packaging: strip VDO_USE_NEXT from sources
Strip out VDO_USE_NEXT and all associated code during upstream
patch generation. Similar to VDO_USE_ALTERNATE, VDO_USE_NEXT
is used to mark code that is too new for us to include in
upstream patches. (Typically this reflects code in linux-next
that we do not expect to be included in the next kernel version.)

Signed-off-by: Matthew Sakai <msakai@redhat.com>